+++ src/app/core/@ngrx/products/products.reducer.ts
@@ -21,13 +21,11 @@ const reducer = createReducer(
       ...state,
       data,
       loading: false,
-      loaded: true,
-      selectedProduct: null
+      loaded: true
     };
   }),
   on(
     ProductsActions.getProductsError,
-    ProductsActions.getProductError,
     (state, { error }) => {
       console.log('GET_PRODUCTS/PRODUCT_ERROR action being handled!');
       return {
@@ -37,26 +35,6 @@ const reducer = createReducer(
         error
       };
     }),
-
-  // GET PRODUCT
-  on(ProductsActions.getProduct, state => {
-    console.log('GET_PRODUCT action being handled!');
-    return {
-      ...state,
-      loading: true,
-      loaded: false
-    };
-  }),
-  on(ProductsActions.getProductSuccess, (state, { product }) => {
-    console.log('GET_PRODUCT action being handled!');
-    const selectedProduct = { ...product };
-    return {
-      ...state,
-      loading: false,
-      loaded: true,
-      selectedProduct
-    };
-  }),
 
   // CREATE PRODUCT
   on(ProductsActions.createProduct, state => {