+++ src/app/products/components/index.ts
@@ -1,3 +1,4 @@
 export { ProductFormComponent } from './product-form/product-form.component';
 export { ProductComponent } from './product/product.component';
 export { ProductListComponent } from './product-list/product-list.component';
+export { ProductViewComponent } from './product-view/product-view.component';