+++ src/app/products/products-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { ProductComponent, ProductListComponent } from './components';
+import { ProductComponent, ProductListComponent, ProductViewComponent } from './components';
 import { ProductsComponent } from './products.component';
 
 const routes: Routes = [
@@ -9,6 +9,10 @@ const routes: Routes = [
     path: 'product-list',
     component: ProductsComponent,
     children: [
+      {
+        path: 'product-view/:productID',
+        component: ProductViewComponent
+      },
       {
         path: 'product/:productID',
         component: ProductComponent
@@ -29,6 +33,7 @@ export class ProductsRoutingModule {
   static components = [
     ProductsComponent,
     ProductListComponent,
-    ProductComponent
+    ProductComponent,
+    ProductViewComponent
   ];
  }